// packages/ext/src/edit/types.ts
export interface CellPatch {
  rowId: string; colId: string; field: string;
  oldValue: unknown; newValue: unknown;
}